+++ packages/frontend/src/pages/welcome.timeline.note.vue
@@ -7,17 +7,17 @@ SPDX-License-Identifier: AGPL-3.0-only
 <div :key="note.id" :class="$style.note">
 	<div class="_panel _gaps_s" :class="$style.content">
 		<div v-if="note.cw != null" :class="$style.richcontent">
-			<div><Mfm :text="note.cw" :author="note.user"/></div>
+			<div><Mfm :text="note.cw" :author="note.user" class="_textAutoSpacing"/></div>
 			<MkCwButton v-model="showContent" :text="note.text" :renote="note.renote" :files="note.files" :poll="note.poll" style="margin: 4px 0;"/>
 			<div v-if="showContent">
 				<MkA v-if="note.replyId" class="reply" :to="`/notes/${note.replyId}`"><i class="ti ti-arrow-back-up"></i></MkA>
-				<Mfm v-if="note.text" :text="note.text" :author="note.user"/>
+				<Mfm v-if="note.text" :text="note.text" :author="note.user" class="_textAutoSpacing"/>
 				<MkA v-if="note.renoteId" class="rp" :to="`/notes/${note.renoteId}`">RN: ...</MkA>
 			</div>
 		</div>
 		<div v-else ref="noteTextEl" :class="[$style.text, { [$style.collapsed]: shouldCollapse }]">
 			<MkA v-if="note.replyId" class="reply" :to="`/notes/${note.replyId}`"><i class="ti ti-arrow-back-up"></i></MkA>
-			<Mfm v-if="note.text" :text="note.text" :author="note.user"/>
+			<Mfm v-if="note.text" :text="note.text" :author="note.user" class="_textAutoSpacing"/>
 			<MkA v-if="note.renoteId" class="rp" :to="`/notes/${note.renoteId}`">RN: ...</MkA>
 		</div>
 		<div v-if="note.files && note.files.length > 0 && (note.cw == null || showContent)" :class="$style.richcontent">